+++ src/typing_ex/typing.cljs
@@ -10,7 +10,7 @@
    [reagent.dom :as rdom]
    [typing-ex.plot :refer [bar-chart]]))
 
-(def ^:private version "1.18.0")
+(def ^:private version "1.18.1-SNAPSHOT")
 (def ^:private timeout 60)
 (def ^:private todays-limit 10)