+++ src/typing_ex/view/page.clj
@@ -10,7 +10,7 @@
    [typing-ex.plot :refer [scatter]]
    [clojure.test :as t]))
 
-(def ^:private version "1.18.0")
+(def ^:private version "1.18.1-SNAPSHOT")
 
 (defn page [& contents]
   [::response/ok
@@ -189,7 +189,7 @@
      [:h2 "Typing: " login " Records"]
      [:p "付け焼き刃はもろい。毎日 10 分 x 3 セット。"]
      [:div.d-inline-flex
-      (when (< 1 (count todays))
+      (when (< 9 (count todays))
         [:div.px-2.mx-auto (scatter 300 150 todays) [:br] [:b "TODAYS"]])
       [:div.px-2]
       [:div.px-2.mx-auto (scatter 300 150 positives) [:br] [:b "TOTAL"]]]